+++ login-workflow/src/screens/ForgotPassword.tsx
@@ -10,23 +10,19 @@ import {
 import { Trans } from 'react-i18next';
 import { useHistory } from 'react-router-dom';
 import { BrandedCardContainer, SimpleDialog, FinishState } from '../components';
-import {
-    CardHeader,
-    Typography,
-    CardContent,
-    Divider,
-    CardActions,
-    Grid,
-    Button,
-    TextField,
-    makeStyles,
-    Theme,
-    createStyles,
-    useTheme,
-} from '@material-ui/core';
+import Button from '@material-ui/core/Button';
+import CardActions from '@material-ui/core/CardActions';
+import CardContent from '@material-ui/core/CardContent';
+import CardHeader from '@material-ui/core/CardHeader';
+import Divider from '@material-ui/core/Divider';
+import Typography from '@material-ui/core/Typography';
+import Grid from '@material-ui/core/Grid';
+import TextField from '@material-ui/core/TextField';
+import { Theme, createStyles, makeStyles, useTheme } from '@material-ui/core/styles';
 import CheckCircle from '@material-ui/icons/CheckCircle';
-import { useDialogStyles } from '../styles';
 import clsx from 'clsx';
+import { sharedDialogStyles } from '../styles';
+const useDialogStyles = makeStyles(sharedDialogStyles);
 
 const useStyles = makeStyles((theme: Theme) =>
     createStyles({